# example/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.NoError:
    type: object
  handler.NoInfo:
    type: object
  handler.PaginationInfo:
    description: pagination info
    properties:
      current:
        description: current page no
        example: 1
        type: integer
      has_next:
        description: whether next page exists
        example: true
        type: boolean
      has_prev:
        description: whether previous page exits
        example: false
        type: boolean
      total:
        description: total resources count
        example: 100
        type: integer
    type: object
  handler.Response-array_handler_User-handler_PaginationInfo-handler_NoError:
    properties:
      data:
        items:
          $ref: '#/definitions/handler.User'
        type: array
      error:
        $ref: '#/definitions/handler.NoError'
      info:
        $ref: '#/definitions/handler.PaginationInfo'
      message:
        type: string
    type: object
  handler.Response-bool-handler_NoInfo-handler_NoError:
    properties:
      data:
        type: boolean
      error:
        $ref: '#/definitions/handler.NoError'
      info:
        $ref: '#/definitions/handler.NoInfo'
      message:
        type: string
    type: object
  handler.Response-handler_User-handler_NoInfo-handler_NoError:
    properties:
      data:
        $ref: '#/definitions/handler.User'
      error:
        $ref: '#/definitions/handler.NoError'
      info:
        $ref: '#/definitions/handler.NoInfo'
      message:
        type: string
    type: object
  handler.User:
    properties:
      created_at:
        example: "2026-02-12T09:04:30.145+00:00"
        type: string
      created_by:
        example: 7e602c5d-8460-4790-b153-a4feb5ceba3a
        type: string
      deleted_at:
        example: "2026-02-12T09:04:30.145+00:00"
        type: string
      deleted_by:
        example: 7e602c5d-8460-4790-b153-a4feb5ceba3a
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 7e602c5d-8460-4790-b153-a4feb5ceba3a
        type: string
      updated_at:
        example: "2026-02-12T09:04:30.145+00:00"
        type: string
      updated_by:
        example: 7e602c5d-8460-4790-b153-a4feb5ceba3a
        type: string
    type: object
  handler.UserFields:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  pagination.SortOrder:
    enum:
    - asc
    - desc
    type: string
    x-enum-varnames:
    - SortOrderAscending
    - SortOrderDescending
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /users/create:
    post:
      consumes:
      - application/json
      description: create an user
      parameters:
      - description: create user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserFields'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-handler_User-handler_NoInfo-handler_NoError'
      summary: create an user
      tags:
      - user
  /users/delete/{id}:
    delete:
      consumes:
      - application/json
      description: delete an user
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-bool-handler_NoInfo-handler_NoError'
      summary: delete an user
      tags:
      - user
  /users/destroy/{id}:
    delete:
      consumes:
      - application/json
      description: destroy an user
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-bool-handler_NoInfo-handler_NoError'
      summary: destroy an user
      tags:
      - user
  /users/list:
    get:
      consumes:
      - application/json
      description: list users
      parameters:
      - description: how many resource you want to fetch per set. default to 10
        in: query
        name: limit
        type: integer
      - description: the page no you want fetch. default to 1.
        in: query
        name: page
        type: integer
      - description: the term to be searched for. default is ""
        in: query
        name: search
        type: string
      - description: the field you want to sort the result set by. defaults to 'created_by'
        in: query
        name: sort_by
        type: string
      - description: the sorting order. validation options are 'asc' | 'desc'. default
          to 'desc'
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
        x-enum-varnames:
        - SortOrderAscending
        - SortOrderDescending
      - description: whether to show deleted resources(only). defaults to false
        in: query
        name: trash
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-array_handler_User-handler_PaginationInfo-handler_NoError'
      summary: list users
      tags:
      - user
  /users/restore/{id}:
    patch:
      consumes:
      - application/json
      description: restore an user
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-bool-handler_NoInfo-handler_NoError'
      summary: restore an user
      tags:
      - user
  /users/update/{id}:
    put:
      consumes:
      - application/json
      description: update an user
      parameters:
      - description: update user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserFields'
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-bool-handler_NoInfo-handler_NoError'
      summary: update an user
      tags:
      - user
  /users/view/{id}:
    get:
      consumes:
      - application/json
      description: view an user
      parameters:
      - description: id of the user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-handler_User-handler_NoInfo-handler_NoError'
      summary: view an user
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"